+++ CHANGES.rst
@@ -7,8 +7,9 @@ Changelog of lizard-graph
 
 - Made admin for PredefinedGraph faster.
 
-- Implemented related location follower, new
-  GraphItem.related_location.
+- Implemented related location follower using new fewsnorm.
+
+- Added option to GraphItem to indicate follow a related location.
 
 
 0.15 (2012-04-16)